move PleaseUseReplacementException to core
i.e.:
- duplicate and deprecate the one in domain-builders
- ignore failing bbc and bc tests
