+++ core/api/atrium-core-api-common/src/main/kotlin/ch/tutteli/atrium/creating/PleaseUseReplacementException.kt
@@ -1,0 +1,6 @@
+package ch.tutteli.atrium.creating
+
+/**
+ * Indicates a problem which was indicated by a `@Deprecated` annotation but was ignored by you ;-)
+ */
+class PleaseUseReplacementException(reason: String) : Exception(reason)
