+++ misc/deprecated/domain/builders/atrium-domain-builders-common/src/main/kotlin/ch/tutteli/atrium/domain/builders/creating/PleaseUseReplacementException.kt
@@ -3,5 +3,13 @@ package ch.tutteli.atrium.domain.builders.creating
 /**
  * Indicates a problem which was indicated by a `@Deprecated` annotation but was ignored by you ;-)
  */
-//TODO move to logic with 0.16.0
-class PleaseUseReplacementException(reason: String) : Exception(reason)
+@Deprecated(
+    "Use PleaseUseReplacementException from core; will be removed with 0.17.0",
+    ReplaceWith("ch.tutteli.atrium.creating.PleaseUseReplacementException")
+)
+class PleaseUseReplacementException
+@Deprecated(
+    "Use PleaseUseReplacementException from core; will be removed with 0.17.0",
+    ReplaceWith("ch.tutteli.atrium.creating.PleaseUseReplacementException(reason)")
+)
+constructor(reason: String) : Exception(reason)
