+++ misc/deprecated/domain/builders/atrium-domain-builders-common/src/main/kotlin/ch/tutteli/atrium/domain/builders/utils/mapArguments.kt
@@ -6,7 +6,7 @@
 package ch.tutteli.atrium.domain.builders.utils
 
 import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.domain.builders.creating.PleaseUseReplacementException
+import ch.tutteli.atrium.creating.PleaseUseReplacementException
 import kotlin.js.JsName
 
 /**
